[fix]: rcv-mcp-v3.0.0
