+++ src/oci-recovery-mcp-server/oracle/oci_recovery_mcp_server/tests/test_auth_and_client_factories.py
@@ -1,0 +1,278 @@
+"""
+Copyright (c) 2025, 2026 Oracle and/or its affiliates.
+Licensed under the Universal Permissive License v1.0 as shown at
+https://oss.oracle.com/licenses/upl.
+
+Credential resolution and OCI client construction: the profile auth context,
+the informational config read, and every client factory.
+"""
+
+from types import SimpleNamespace
+from unittest.mock import MagicMock, patch
+
+from fastmcp import Client
+
+import oracle.oci_recovery_mcp_server.server as server
+
+
+class TestProfileClientFactories:
+    """Client factories now resolve apikey/session credentials through
+    oracle_mcp_common.build_auth_context() (see server._build_profile_auth_context),
+    so these assert the config/signer handed to the SDK constructor.
+    """
+
+    @patch(
+        "oracle.oci_recovery_mcp_server.server._wrap_oci_client",
+        side_effect=lambda client, **_: client,
+    )
+    @patch("oracle.oci_recovery_mcp_server.server.oci.recovery.DatabaseRecoveryClient")
+    @patch("oracle.oci_recovery_mcp_server.server._build_profile_auth_context")
+    def test_get_recovery_client_apikey_uses_profile_auth_context(
+        self,
+        mock_auth_context,
+        mock_client,
+        _mock_wrap,
+    ):
+        signer = object()
+        mock_auth_context.return_value = SimpleNamespace(
+            config={"region": "us-ashburn-1", "tenancy": "ocid1.tenancy.oc1..t"},
+            signer=signer,
+        )
+
+        result = server.get_recovery_client(region="us-phoenix-1", request_id="rid")
+
+        args, kwargs = mock_client.call_args
+        assert args[0]["region"] == "us-phoenix-1"
+        assert args[0]["tenancy"] == "ocid1.tenancy.oc1..t"
+        assert args[0]["additional_user_agent"].startswith(server._USER_AGENT_NAME)
+        assert kwargs["signer"] is signer
+        assert result is mock_client.return_value
+
+    @patch(
+        "oracle.oci_recovery_mcp_server.server._wrap_oci_client",
+        side_effect=lambda client, **_: client,
+    )
+    @patch("oracle.oci_recovery_mcp_server.server.oci.monitoring.MonitoringClient")
+    @patch("oracle.oci_recovery_mcp_server.server._build_profile_auth_context")
+    def test_get_monitoring_client_session_uses_profile_auth_context(
+        self,
+        mock_auth_context,
+        mock_client,
+        _mock_wrap,
+    ):
+        signer = object()
+        mock_auth_context.return_value = SimpleNamespace(
+            config={"region": "us-ashburn-1"}, signer=signer
+        )
+
+        result = server.get_monitoring_client(region="us-phoenix-1", request_id="rid")
+
+        args, kwargs = mock_client.call_args
+        assert args[0]["region"] == "us-phoenix-1"
+        assert kwargs["signer"] is signer
+        assert result is mock_client.return_value
+
+    @patch(
+        "oracle.oci_recovery_mcp_server.server._wrap_oci_client",
+        side_effect=lambda client, **_: client,
+    )
+    @patch("oracle.oci_recovery_mcp_server.server.oci.recovery.DatabaseRecoveryClient")
+    @patch("oracle.oci_recovery_mcp_server.server._http_config_and_signer")
+    @patch("oracle.oci_recovery_mcp_server.server._serving_http", return_value=True)
+    def test_http_client_uses_request_scoped_token_exchange_signer(
+        self,
+        _mock_serving_http,
+        mock_http_auth,
+        mock_client,
+        _mock_wrap,
+    ):
+        first, second = object(), object()
+        mock_http_auth.side_effect = [
+            ({"region": "us-phoenix-1"}, first),
+            ({"region": "us-phoenix-1"}, second),
+        ]
+
+        server.get_recovery_client(region="us-phoenix-1", request_id="rid")
+        server.get_recovery_client(region="us-phoenix-1", request_id="rid")
+
+        # A fresh signer per call: UPST signers are never cached.
+        assert [call.kwargs["signer"] for call in mock_client.call_args_list] == [
+            first,
+            second,
+        ]
+        assert mock_client.call_args_list[0].args[0]["region"] == "us-phoenix-1"
+
+
+def test_informational_config_reads_same_file_as_the_credentials(monkeypatch, tmp_path):
+    """The OCI SDK only falls back to OCI_CONFIG_FILE when ~/.oci/config is absent,
+    while oracle_mcp_common.resolve_config_file() always prefers it. Tenancy and
+    region lookups must follow the credentials, not the default path.
+    """
+    key_file = tmp_path / "api_key.pem"
+    key_file.write_text("not-a-real-key", encoding="utf-8")
+    alt_config = tmp_path / "alt_config"
+    alt_config.write_text(
+        "[PROFILE1]\n"
+        "tenancy=ocid1.tenancy.oc1..alt\n"
+        "region=eu-frankfurt-1\n"
+        "user=ocid1.user.oc1..alt\n"
+        "fingerprint=aa:bb\n"
+        f"key_file={key_file}\n",
+        encoding="utf-8",
+    )
+    monkeypatch.setenv("OCI_CONFIG_FILE", str(alt_config))
+    monkeypatch.setenv("ORACLE_MCP_AUTH_PROFILE", "PROFILE1")
+    monkeypatch.setenv("ORACLE_MCP_AUTH_METHOD", "apikey")
+    monkeypatch.delenv("TENANCY_ID_OVERRIDE", raising=False)
+    monkeypatch.delenv("ORACLE_MCP_TENANCY_ID", raising=False)
+
+    config = server._load_oci_config_for_server()
+
+    assert config["tenancy"] == "ocid1.tenancy.oc1..alt"
+    assert config["additional_user_agent"].startswith(server._USER_AGENT_NAME)
+    assert server.get_tenancy() == "ocid1.tenancy.oc1..alt"
+    assert server._effective_region() == "eu-frankfurt-1"
+
+
+def test_profile_auth_context_defers_to_the_shared_library(monkeypatch):
+    captured = []
+
+    def fake_build_auth_context(*args):
+        captured.append(args)
+        return SimpleNamespace(config={"region": "home"}, signer=object())
+
+    monkeypatch.setattr(server, "build_auth_context", fake_build_auth_context)
+    monkeypatch.setenv("ORACLE_MCP_AUTH_PROFILE", "PROFILE1")
+    for name in server._CANONICAL_AUTH_TYPE_ENV:
+        monkeypatch.delenv(name, raising=False)
+
+    # Only the deprecated unseparated spelling is translated.
+    monkeypatch.setenv("ORACLE_MCP_AUTH_METHOD", "apikey")
+    server._build_profile_auth_context()
+    # Everything else -- including an unset value, which means auto-detect -- is
+    # passed through so the library resolves type and profile itself.
+    monkeypatch.setenv("ORACLE_MCP_AUTH_METHOD", "session")
+    server._build_profile_auth_context()
+    monkeypatch.delenv("ORACLE_MCP_AUTH_METHOD", raising=False)
+    server._build_profile_auth_context()
+
+    assert captured[0][0].auth_type == server.AuthType.API_KEY
+    assert captured[1] == ()
+    assert captured[2] == ()
+
+
+def test_client_factories_use_profile_and_http_auth_paths(monkeypatch):
+    monkeypatch.setattr(
+        server,
+        "_wrap_oci_client",
+        lambda client, **kwargs: (client, kwargs["client_name"]),
+    )
+
+    recovery_client = MagicMock(return_value="recovery-client")
+    database_client = MagicMock(return_value="database-client")
+    identity_client = MagicMock(return_value="identity-client")
+    monitoring_client = MagicMock(return_value="monitoring-client")
+    monkeypatch.setattr(server.oci.recovery, "DatabaseRecoveryClient", recovery_client)
+    monkeypatch.setattr(server.oci.database, "DatabaseClient", database_client)
+    monkeypatch.setattr(server.oci.identity, "IdentityClient", identity_client)
+    monkeypatch.setattr(server.oci.monitoring, "MonitoringClient", monitoring_client)
+
+    # apikey/session: config + signer come from the shared auth context.
+    profile_signer = object()
+    monkeypatch.setattr(server, "_serving_http", lambda: False)
+    monkeypatch.setattr(
+        server,
+        "_build_profile_auth_context",
+        lambda: SimpleNamespace(config={"region": "home"}, signer=profile_signer),
+    )
+
+    assert server.get_recovery_client(region="us-ashburn-1")[1] == "recovery"
+    assert server.get_database_client(region="us-chicago-1")[1] == "database"
+    assert server.get_identity_client()[1] == "identity"
+    assert server.get_monitoring_client(region="us-ashburn-1")[1] == "monitoring"
+    assert recovery_client.call_args.args[0]["region"] == "us-ashburn-1"
+    assert database_client.call_args.args[0]["region"] == "us-chicago-1"
+    # get_identity_client passes no region, so the profile's home region stands.
+    assert identity_client.call_args.args[0]["region"] == "home"
+    for client in (recovery_client, database_client, identity_client, monitoring_client):
+        assert client.call_args.kwargs["signer"] is profile_signer
+
+    # HTTP: regional config + per-request token-exchange signer.
+    http_signer = object()
+    monkeypatch.setattr(server, "_serving_http", lambda: True)
+    monkeypatch.setenv("OCI_REGION", "us-ashburn-1")
+    monkeypatch.setattr(
+        server,
+        "_http_config_and_signer",
+        lambda region=None: ({"region": region or "us-ashburn-1"}, http_signer),
+    )
+
+    assert server.get_recovery_client(region="us-phoenix-1")[1] == "recovery"
+    assert server.get_database_client()[1] == "database"
+    assert recovery_client.call_args.args[0]["region"] == "us-phoenix-1"
+    assert database_client.call_args.args[0]["region"] == "us-ashburn-1"
+    assert recovery_client.call_args.kwargs["signer"] is http_signer
+    assert database_client.call_args.kwargs["signer"] is http_signer
+
+
+def test_limits_work_request_and_subscription_client_factories(monkeypatch):
+    profile_signer = object()
+    monkeypatch.setattr(server, "_serving_http", lambda: False)
+    monkeypatch.setattr(
+        server,
+        "_build_profile_auth_context",
+        lambda: SimpleNamespace(config={"region": "home-region"}, signer=profile_signer),
+    )
+    monkeypatch.setattr(
+        server,
+        "_wrap_oci_client",
+        lambda client, **kwargs: (client, kwargs["client_name"]),
+    )
+
+    limits_client = MagicMock(return_value="limits-client")
+    work_request_client = MagicMock(return_value="work-request-client")
+    subscribed_service_client = MagicMock(return_value="subscription-client")
+    monkeypatch.setattr(server.oci.limits, "LimitsClient", limits_client)
+    monkeypatch.setattr(
+        server.oci.work_requests, "WorkRequestClient", work_request_client
+    )
+    monkeypatch.setattr(
+        server.oci.onesubscription,
+        "SubscribedServiceClient",
+        subscribed_service_client,
+    )
+
+    assert server.get_limits_client(region="us-phoenix-1")[1] == "limits"
+    assert (
+        server.get_work_request_client(region="us-chicago-1")[1] == "work_requests"
+    )
+    assert (
+        server.get_onesubscription_client(region="us-ashburn-1")[1]
+        == "onesubscription"
+    )
+    assert limits_client.call_args.args[0]["region"] == "us-phoenix-1"
+    assert work_request_client.call_args.args[0]["region"] == "us-chicago-1"
+    assert subscribed_service_client.call_args.args[0]["region"] == "us-ashburn-1"
+    assert limits_client.call_args.kwargs["signer"] is profile_signer
+
+    # No explicit region falls back to the resolved profile's home region.
+    assert server.get_limits_client()[1] == "limits"
+    assert limits_client.call_args.args[0]["region"] == "home-region"
+
+    http_signer = object()
+    monkeypatch.setattr(server, "_serving_http", lambda: True)
+    monkeypatch.setenv("OCI_REGION", "us-ashburn-1")
+    monkeypatch.setattr(
+        server,
+        "_http_config_and_signer",
+        lambda region=None: ({"region": region or "us-ashburn-1"}, http_signer),
+    )
+    assert server.get_work_request_client(region="us-sanjose-1")[1] == "work_requests"
+    assert server.get_limits_client(region="us-sanjose-1")[1] == "limits"
+    assert (
+        server.get_onesubscription_client(region="us-sanjose-1")[1]
+        == "onesubscription"
+    )
+    assert work_request_client.call_args.kwargs["signer"] is http_signer
+    assert limits_client.call_args.kwargs["signer"] is http_signer
+    assert subscribed_service_client.call_args.kwargs["signer"] is http_signer
