+++ src/oci-recovery-mcp-server/oracle/oci_recovery_mcp_server/tests/test_database_service_tools.py
@@ -1,0 +1,626 @@
+"""
+Copyright (c) 2025, 2026 Oracle and/or its affiliates.
+Licensed under the Universal Permissive License v1.0 as shown at
+https://oss.oracle.com/licenses/upl.
+
+Database Service tools: databases, DB homes, DB systems, and their
+compartment-subtree and error paths.
+"""
+
+from types import SimpleNamespace
+from unittest.mock import MagicMock
+
+import pytest
+
+from _helpers import _response
+import oracle.oci_recovery_mcp_server.models as models
+import oracle.oci_recovery_mcp_server.server as server
+
+
+def test_database_tools_resolve_compartment_paths_and_enrich_backups(monkeypatch):
+    db_client = MagicMock()
+    recovery_client = MagicMock()
+    monkeypatch.setattr(
+        models.oci.util,
+        "to_dict",
+        lambda obj: obj if isinstance(obj, dict) else getattr(obj, "__dict__", obj),
+    )
+    monkeypatch.setattr(
+        server,
+        "get_database_client",
+        lambda region=None, request_id=None: db_client,
+    )
+    monkeypatch.setattr(
+        server,
+        "get_recovery_client",
+        lambda region=None, request_id=None: recovery_client,
+    )
+    monkeypatch.setattr(
+        server,
+        "_resolve_compartment_id",
+        lambda compartment_id, **_kwargs: compartment_id or "tenancy",
+    )
+    monkeypatch.setattr(
+        server,
+        "_fetch_db_home_ids_for_compartment",
+        lambda compartment_id, region=None: ["home1"],
+    )
+
+    recovery_client.list_protected_databases.return_value = _response(
+        SimpleNamespace(items=[{"databaseId": "db1", "protectionPolicyId": "policy1"}])
+    )
+    db_client.list_databases.return_value = _response(
+        SimpleNamespace(items=[SimpleNamespace(id="db1", db_name="DB1")])
+    )
+    db_client.get_database.return_value = _response(
+        {
+            "id": "db1",
+            "compartmentId": "compartment",
+            "dbBackupConfig": {"isAutoBackupEnabled": True},
+        }
+    )
+
+    databases = server.list_databases(
+        compartment_id="compartment",
+        system_id="system1",
+        limit=10,
+        page="page",
+        sort_by="DBNAME",
+        sort_order="ASC",
+        lifecycle_state="AVAILABLE",
+        db_name="DB1",
+        region="us-ashburn-1",
+    )
+    assert databases[0].id == "db1"
+    assert databases[0].db_backup_config.is_auto_backup_enabled is True
+    assert databases[0].protection_policy_id == "policy1"
+    assert db_client.list_databases.call_args.kwargs["db_home_id"] == "home1"
+
+    db_client.get_database.return_value = _response(
+        {"id": "db1", "compartmentId": "compartment"}
+    )
+    recovery_client.list_protected_databases.return_value = _response(
+        [{"databaseId": "db1", "protectionPolicyId": "policy2"}]
+    )
+    database = server.get_database("db1", region="us-ashburn-1")
+    assert database.id == "db1"
+    assert database.protection_policy_id == "policy2"
+
+    db_client.list_databases.return_value = _response(
+        SimpleNamespace(
+            items=[
+                {
+                    "id": "db1",
+                    "dbUniqueName": "DB1_UNQ",
+                    "dbBackupConfig": {"isAutoBackupEnabled": True},
+                }
+            ]
+        )
+    )
+    db_client.list_backups.side_effect = [
+        _response(
+            SimpleNamespace(
+                items=[
+                    SimpleNamespace(
+                        id="backup1",
+                        database_id="db1",
+                        backup_destination_type="DBRS",
+                    )
+                ]
+            ),
+            has_next_page=True,
+            next_page="backup-page-2",
+        ),
+        _response(
+            SimpleNamespace(
+                items=[
+                    {
+                        "id": "backup2",
+                        "databaseId": "db1",
+                        "databaseSizeInGbs": 11,
+                    }
+                ]
+            )
+        ),
+    ]
+
+    backups = server.list_backups(
+        compartment_id="compartment",
+        lifecycle_state="ACTIVE",
+        type="FULL",
+        region="us-ashburn-1",
+    )
+    assert [backup["id"] for backup in backups] == ["backup1", "backup2"]
+    assert backups[0]["db_unique_name"] == "DB1_UNQ"
+    assert backups[1]["database-size-in-gbs"] == 11
+    assert db_client.list_backups.call_args_list[1].kwargs["page"] == "backup-page-2"
+
+    db_client.get_backup.return_value = _response(
+        SimpleNamespace(
+            id="backup1",
+            database_id="db1",
+            database_size_in_gbs=8,
+            retention_period_in_days=30,
+        )
+    )
+    db_client.get_database.return_value = _response(
+        {
+            "dbUniqueName": "DB1_UNQ",
+            "dbBackupConfig": {
+                "backupDestinationDetails": [{"type": "RECOVERY_SERVICE"}]
+            },
+        }
+    )
+    backup = server.get_backup("backup1", region="us-ashburn-1")
+    assert backup["database-size-in-gbs"] == 8
+    assert backup["backup-destination-type"] == "DBRS"
+    assert backup["db_unique_name"] == "DB1_UNQ"
+
+    db_client.list_databases.return_value = _response(
+        [
+            {
+                "id": "db1",
+                "dbName": "DB1",
+                "dbBackupConfig": {
+                    "isAutoBackupEnabled": True,
+                    "backupDestinationDetails": [
+                        {"type": "RECOVERY_SERVICE", "id": "dest1"}
+                    ],
+                },
+            },
+            {"id": "db2", "dbName": "DB2"},
+        ]
+    )
+    db_client.get_database.return_value = _response(
+        {
+            "id": "db2",
+            "dbName": "DB2",
+            "dbBackupConfig": {"isAutoBackupEnabled": False},
+        }
+    )
+    db_client.list_backups.side_effect = None
+    db_client.list_backups.return_value = _response(
+        [SimpleNamespace(time_ended="2024-01-02T00:00:00Z")]
+    )
+    summary = server.summarize_protected_database_backup_destination(
+        compartment_id="compartment",
+        region="us-ashburn-1",
+        include_last_backup_time=True,
+        db_name="DB",
+        limit_per_home=50,
+        max_db_homes=1,
+        max_total_databases=2,
+    )
+    assert summary.total_databases == 2
+    assert summary.counts_by_destination_type == {"DBRS": 1}
+    assert summary.unconfigured_count == 1
+    assert [item.database_id for item in summary.items] == ["db1", "db2"]
+    assert summary.items[0].last_backup_time.isoformat() == "2024-01-02T00:00:00+00:00"
+
+
+def test_database_child_scope_tools_deduplicate_results(monkeypatch):
+    db_client = MagicMock()
+    recovery_client = MagicMock()
+    monkeypatch.setattr(
+        models.oci.util,
+        "to_dict",
+        lambda obj: obj if isinstance(obj, dict) else getattr(obj, "__dict__", obj),
+    )
+    monkeypatch.setattr(
+        server,
+        "get_database_client",
+        lambda region=None, request_id=None: db_client,
+    )
+    monkeypatch.setattr(
+        server,
+        "get_recovery_client",
+        lambda region=None, request_id=None: recovery_client,
+    )
+    monkeypatch.setattr(
+        server, "_resolve_compartment_id", lambda value, **_kwargs: value
+    )
+    monkeypatch.setattr(
+        server,
+        "_compartment_ids_for_tool",
+        lambda compartment_id, fetch_for_child_compartment, request_id=None: [
+            "compartment-a",
+            "compartment-b",
+        ],
+    )
+    monkeypatch.setattr(
+        server,
+        "_fetch_db_home_ids_for_compartment",
+        lambda compartment_id, region=None: ["home1"],
+    )
+    recovery_client.list_protected_databases.return_value = _response([])
+
+    db_client.list_databases.side_effect = [
+        _response(
+            SimpleNamespace(
+                items=[
+                    {
+                        "id": "db1",
+                        "dbName": "DB1",
+                        "dbBackupConfig": {"isAutoBackupEnabled": True},
+                    }
+                ]
+            )
+        ),
+        _response(
+            SimpleNamespace(
+                items=[
+                    {
+                        "id": "db1",
+                        "dbName": "DB1 Duplicate",
+                        "dbBackupConfig": {"isAutoBackupEnabled": True},
+                    },
+                    {
+                        "id": "db2",
+                        "dbName": "DB2",
+                        "dbBackupConfig": {"isAutoBackupEnabled": True},
+                    },
+                ]
+            )
+        ),
+    ]
+    databases = server.list_databases(
+        compartment_id="root",
+        fetch_for_child_compartment=True,
+    )
+    assert [database.id for database in databases] == ["db1", "db2"]
+
+    db_client.reset_mock()
+    db_client.list_databases.side_effect = [
+        _response(
+            SimpleNamespace(
+                items=[
+                    {
+                        "id": "db1",
+                        "dbUniqueName": "DB1_UNQ",
+                        "dbBackupConfig": {"isAutoBackupEnabled": True},
+                    }
+                ]
+            )
+        ),
+        _response(
+            SimpleNamespace(
+                items=[
+                    {
+                        "id": "db2",
+                        "dbUniqueName": "DB2_UNQ",
+                        "dbBackupConfig": {"isAutoBackupEnabled": True},
+                    }
+                ]
+            )
+        ),
+    ]
+    db_client.list_backups.side_effect = [
+        _response(SimpleNamespace(items=[{"id": "backup1", "databaseId": "db1"}])),
+        _response(
+            SimpleNamespace(
+                items=[
+                    {"id": "backup1", "databaseId": "db2"},
+                    {"id": "backup2", "databaseId": "db2"},
+                ]
+            )
+        ),
+    ]
+    backups = server.list_backups(
+        compartment_id="root",
+        fetch_for_child_compartment=True,
+    )
+    assert [backup["id"] for backup in backups] == ["backup1", "backup2"]
+
+    db_client.reset_mock()
+    db_client.list_databases.side_effect = [
+        _response(
+            [
+                {
+                    "id": "db1",
+                    "dbName": "DB1",
+                    "dbBackupConfig": {
+                        "isAutoBackupEnabled": True,
+                        "backupDestinationDetails": [
+                            {"type": "RECOVERY_SERVICE", "id": "dest1"}
+                        ],
+                    },
+                }
+            ]
+        ),
+        _response(
+            [
+                {
+                    "id": "db1",
+                    "dbName": "DB1 Duplicate",
+                    "dbBackupConfig": {
+                        "isAutoBackupEnabled": True,
+                        "backupDestinationDetails": [
+                            {"type": "RECOVERY_SERVICE", "id": "dest1"}
+                        ],
+                    },
+                },
+                {
+                    "id": "db2",
+                    "dbName": "DB2",
+                    "dbBackupConfig": {"isAutoBackupEnabled": False},
+                },
+            ]
+        ),
+    ]
+    summary = server.summarize_protected_database_backup_destination(
+        compartment_id="root",
+        fetch_for_child_compartment=True,
+        db_home_id="home1",
+        include_last_backup_time=False,
+    )
+    assert [item.database_id for item in summary.items] == ["db1", "db2"]
+    assert summary.total_databases == 3
+
+    db_client.reset_mock()
+    db_client.list_db_homes.side_effect = [
+        _response([SimpleNamespace(id="home1")]),
+        _response([SimpleNamespace(id="home1"), SimpleNamespace(id="home2")]),
+    ]
+    homes = server.list_db_homes(
+        compartment_id="root",
+        fetch_for_child_compartment=True,
+    )
+    assert [home.id for home in homes] == ["home1", "home2"]
+
+    db_client.reset_mock()
+    db_client.list_db_systems.side_effect = [
+        _response([SimpleNamespace(id="system1")]),
+        _response([SimpleNamespace(id="system1"), SimpleNamespace(id="system2")]),
+    ]
+    systems = server.list_db_systems(
+        compartment_id="root",
+        fetch_for_child_compartment=True,
+    )
+    assert [system.id for system in systems] == ["system1", "system2"]
+
+
+def test_database_home_and_system_tools_apply_pagination_and_defaults(monkeypatch):
+    db_client = MagicMock()
+    monkeypatch.setattr(
+        models.oci.util,
+        "to_dict",
+        lambda obj: obj if isinstance(obj, dict) else getattr(obj, "__dict__", obj),
+    )
+    monkeypatch.setattr(
+        server,
+        "get_database_client",
+        lambda region=None, request_id=None: db_client,
+    )
+    monkeypatch.setattr(server, "get_tenancy", lambda: "tenancy")
+    monkeypatch.setattr(
+        server,
+        "_resolve_compartment_id",
+        lambda compartment_id, **_kwargs: f"resolved-{compartment_id}",
+    )
+
+    db_client.list_db_homes.side_effect = [
+        _response(
+            SimpleNamespace(items=[SimpleNamespace(id="home1", display_name="Home 1")]),
+            has_next_page=True,
+            next_page="home-page-2",
+        ),
+        _response([SimpleNamespace(id="home2", display_name="Home 2")]),
+    ]
+    homes = server.list_db_homes(
+        compartment_id=None,
+        db_system_id=None,
+        limit=1,
+        page="home-page-1",
+        region="us-ashburn-1",
+    )
+    assert [home.id for home in homes] == ["home1", "home2"]
+    assert (
+        db_client.list_db_homes.call_args_list[0].kwargs["compartment_id"]
+        == "resolved-tenancy"
+    )
+    assert db_client.list_db_homes.call_args_list[1].kwargs["page"] == "home-page-2"
+
+    db_client.get_db_home.return_value = _response(
+        SimpleNamespace(id="home1", display_name="Home 1")
+    )
+    assert server.get_db_home("home1", region="us-ashburn-1").id == "home1"
+
+    db_client.list_db_systems.side_effect = [
+        _response(
+            [SimpleNamespace(id="system1", display_name="System 1")],
+            has_next_page=True,
+            next_page="system-page-2",
+        ),
+        _response(SimpleNamespace(items=[SimpleNamespace(id="system2")])),
+    ]
+    systems = server.list_db_systems(
+        compartment_id="compartment",
+        lifecycle_state="AVAILABLE",
+        limit=1,
+        page="system-page-1",
+        region="us-ashburn-1",
+    )
+    assert [system.id for system in systems] == ["system1", "system2"]
+    assert (
+        db_client.list_db_systems.call_args_list[0].kwargs["lifecycle_state"]
+        == "AVAILABLE"
+    )
+    assert db_client.list_db_systems.call_args_list[1].kwargs["page"] == "system-page-2"
+
+    db_client.get_db_system.return_value = _response(
+        SimpleNamespace(id="system1", display_name="System 1")
+    )
+    assert server.get_db_system("system1", region="us-ashburn-1").id == "system1"
+
+
+def test_database_list_branches_and_tool_error_paths(monkeypatch):
+    db_client = MagicMock()
+    recovery_client = MagicMock()
+    monkeypatch.setattr(
+        models.oci.util,
+        "to_dict",
+        lambda obj: obj if isinstance(obj, dict) else getattr(obj, "__dict__", obj),
+    )
+    monkeypatch.setattr(
+        server,
+        "get_database_client",
+        lambda region=None, request_id=None: db_client,
+    )
+    monkeypatch.setattr(
+        server,
+        "get_recovery_client",
+        lambda region=None, request_id=None: recovery_client,
+    )
+    monkeypatch.setattr(
+        server,
+        "_resolve_compartment_id",
+        lambda compartment_id, **_kwargs: compartment_id or "tenancy",
+    )
+    recovery_client.list_protected_databases.return_value = _response([])
+
+    with pytest.raises(ValueError, match="Either db_home_id"):
+        server.list_databases()
+
+    monkeypatch.setattr(
+        server,
+        "_fetch_db_home_ids_for_compartment",
+        lambda compartment_id, region=None: [],
+    )
+    assert server.list_databases(compartment_id="compartment") == []
+
+    monkeypatch.setattr(
+        server,
+        "_fetch_db_home_ids_for_compartment",
+        lambda compartment_id, region=None: ["home1"],
+    )
+    recovery_client.list_protected_databases.side_effect = RuntimeError(
+        "recovery unavailable"
+    )
+    db_client.list_databases.return_value = _response(
+        [
+            {
+                "id": "db1",
+                "dbName": "DB1",
+                "dbBackupConfig": {"isAutoBackupEnabled": True},
+            }
+        ]
+    )
+    databases = server.list_databases(compartment_id="compartment")
+    assert databases[0].protection_policy_id is None
+
+    db_client.list_databases.return_value = _response([{"id": "db2", "dbName": "DB2"}])
+    db_client.get_database.side_effect = RuntimeError("backup config unavailable")
+    databases = server.list_databases(compartment_id="compartment", db_home_id="home1")
+    assert databases[0].id == "db2"
+    assert databases[0].db_backup_config is None
+
+    db_client.reset_mock()
+    db_client.get_database.side_effect = RuntimeError("backup config unavailable")
+    databases = server.list_databases(db_home_id="home1")
+    assert databases[0].id == "db2"
+    assert db_client.list_databases.call_args.kwargs == {"db_home_id": "home1"}
+
+    error_cases = [
+        (
+            server.list_protection_policies,
+            {"compartment_id": "compartment"},
+            "list_protection_policies",
+        ),
+        (
+            server.get_protection_policy,
+            {"protection_policy_id": "policy1"},
+            "get_protection_policy",
+        ),
+        (
+            server.list_recovery_service_subnets,
+            {"compartment_id": "compartment"},
+            "list_recovery_service_subnets",
+        ),
+        (
+            server.get_recovery_service_subnet,
+            {"recovery_service_subnet_id": "rss1"},
+            "get_recovery_service_subnet",
+        ),
+    ]
+    for tool, kwargs, method_name in error_cases:
+        getattr(recovery_client, method_name).side_effect = RuntimeError(
+            f"{method_name} failed"
+        )
+        with pytest.raises(RuntimeError, match=f"{method_name} failed"):
+            tool(**kwargs)
+        getattr(recovery_client, method_name).side_effect = None
+
+    db_error_cases = [
+        (server.get_database, {"database_id": "db1"}, "get_database"),
+        (server.get_backup, {"backup_id": "backup1"}, "get_backup"),
+        (server.list_db_homes, {"compartment_id": "compartment"}, "list_db_homes"),
+        (server.get_db_home, {"db_home_id": "home1"}, "get_db_home"),
+        (server.list_db_systems, {"compartment_id": "compartment"}, "list_db_systems"),
+        (server.get_db_system, {"db_system_id": "system1"}, "get_db_system"),
+    ]
+    for tool, kwargs, method_name in db_error_cases:
+        getattr(db_client, method_name).side_effect = RuntimeError(
+            f"{method_name} failed"
+        )
+        with pytest.raises(RuntimeError, match=f"{method_name} failed"):
+            tool(**kwargs)
+        getattr(db_client, method_name).side_effect = None
+
+    # Guidance tools return the prompt text directly so clients without prompt
+    # support can call them as ordinary tools.
+    assert (
+        server.oci_recovery_service_dashboard_prompt()
+        == server.OCI_RECOVERY_SERVICE_DASHBOARD_PROMPT
+    )
+    assert (
+        server.onboard_database_to_recovery_service()
+        == server.ONBOARD_DATABASE_TO_RECOVERY_SERVICE_PROMPT
+    )
+    assert (
+        server.diagnose_recovery_service_issue()
+        == server.DIAGNOSE_RECOVERY_SERVICE_ISSUE_PROMPT
+    )
+
+
+def test_policy_correlation_survives_an_unreadable_compartment(monkeypatch):
+    # A caller who cannot list protected databases in one compartment of a subtree
+    # gets a 404 there. Discarding the whole correlation would strip the policy
+    # link off every database in every readable compartment too, which reads as
+    # "no protection policy" rather than "could not check".
+    recovery_client = MagicMock()
+    database_client = MagicMock()
+
+    def list_protected_databases(compartment_id=None, page=None):
+        if compartment_id == "denied":
+            raise RuntimeError("NotAuthorizedOrNotFound")
+        return _response(
+            SimpleNamespace(
+                items=[{"databaseId": "db-in-" + compartment_id, "protectionPolicyId": "pol-" + compartment_id}]
+            )
+        )
+
+    recovery_client.list_protected_databases.side_effect = list_protected_databases
+    database_client.list_db_homes.return_value = _response([SimpleNamespace(id="home1")])
+    database_client.list_databases.return_value = _response(
+        [
+            SimpleNamespace(id="db-in-readable", db_name="OPEN", lifecycle_state="AVAILABLE"),
+            SimpleNamespace(id="db-in-denied", db_name="HIDDEN", lifecycle_state="AVAILABLE"),
+        ]
+    )
+
+    monkeypatch.setattr(server, "get_recovery_client", lambda *a, **k: recovery_client)
+    monkeypatch.setattr(server, "get_database_client", lambda *a, **k: database_client)
+    monkeypatch.setattr(
+        server, "_compartment_ids_for_tool", lambda cid, **k: ["readable", "denied"]
+    )
+    monkeypatch.setattr(server, "_resolve_compartment_id", lambda c, **k: c)
+
+    databases = server.list_databases(
+        compartment_id="root", fetch_for_child_compartment=True
+    )
+
+    policies = {d.id: d.protection_policy_id for d in databases}
+    # The readable compartment's correlation survives the other one's failure.
+    assert policies.get("db-in-readable") == "pol-readable"
+    # And the unreadable one is simply unlinked, not fabricated.
+    assert policies.get("db-in-denied") is None
