+++ src/oci-recovery-mcp-server/pyproject.toml
@@ -10,11 +10,10 @@ authors = [
     {name = "Oracle MCP", email = "237432095+oracle-mcp@users.noreply.github.com"},
 ]
 dependencies = [
+    "cryptography==50.0.1",
     "fastmcp==3.4.5",
-    "oci==2.182.1",
+    "oci==2.185.1",
     "pydantic==2.13.4",
-    "mcp>=1.27.0",
-    "pytest-cov>=7.0.0",
     "python-dotenv>=1.0.0",
     "oracle-mcp-common>=0.1.2,<0.2.0",
 ]
@@ -68,6 +67,9 @@ exclude = [
   "/.roo/**",
   "/.cursor/**",
   "/mcp_settings*.json",
+  # Local-only live-integration suite: needs real credentials, is never part of a
+  # release, and its run logs contain tenancy data.
+  "/integration_tests/**",
   "/Containerfile",
   "/.containerignore",
 ]
@@ -77,16 +79,23 @@ dev = [
     "black>=26.3.1",
     "pytest>=9.0.3",
     "pytest-asyncio>=1.2.0",
+    "pytest-cov>=7.0.0",
 ]
+
+[tool.pytest.ini_options]
+testpaths = ["oracle/oci_recovery_mcp_server/tests"]
+norecursedirs = ["integration_tests"]
 
 [tool.coverage.run]
 omit = [
     "**/__init__.py",
     "**/tests/*",
+    "**/integration_tests/*",
     "dist/*",
     ".venv/*",
 ]
 
 [tool.coverage.report]
+omit = ["**/integration_tests/*"]
 precision = 2
 fail_under = 90
